Made class BloomHash that derives from Hash. Removed Factor
