+++ BloomFilter/Hash.cs
@@ -11,53 +11,15 @@ namespace BloomFilterCore
 	public class Hash : IDisposable
 	{
 		public int TableSize = 256;	// Because we are using bytes
-		public bool IsDisposed { get; private set; }
+		public bool IsDisposed { get; protected set; }
 	
-		private static int maxIndex;
-		private static int quantityIndexBytes;
-
 		private byte i;
 		private byte j;
 		private byte l;
 		private byte[] _table;
 		private byte k { get { unchecked { l = (byte)(_table[i] + _table[j]); return _table[l]; } } }
 
-		private string token;		
-		private int _tokCntr = 0;
-		private int tokenCounter { get { return ++_tokCntr; } }
-
-		private int _tokIndx = -1;
-		private int tokenIndex
-		{
-			get
-			{
-				_tokIndx += 1;
-				if (_tokIndx > token.Length - 1)
-				{
-					_tokIndx = 0;
-				}
-				return _tokIndx;
-			}
-		}
-				
-		public Hash(string input, int indexByteSize, int maximumIndex)
-		{
-			token = input;
-			maxIndex = maximumIndex;
-			quantityIndexBytes = indexByteSize;
-			Initialize();
-
-			int prime = 1;
-			byte[] inputBytes = Encoding.ASCII.GetBytes(input);
-			foreach (byte bite in inputBytes)
-			{
-				prime = (prime * 2) + 1;
-				Shuffle(bite * prime);
-			}
-			//Shuffle(TableSize); //Shuffle(input.Length);
-		}
-
-		private void Initialize()
+		public Hash()
 		{
 			Clear();
 			int counter = TableSize;
@@ -67,24 +29,9 @@ namespace BloomFilterCore
 				result.Add((byte)(255 - counter));
 			}
 			_table = result.ToArray();
-			IsDisposed = false;
+			IsDisposed = false;		
 		}
-
-		public IEnumerable<int> GetIndices()
-		{
-			int shiftAmmount = 0;
-			List<byte> bytes;
-			while (!IsDisposed)
-			{
-				shiftAmmount = token[tokenIndex] * tokenCounter;
-				Shuffle(shiftAmmount); // Shuffle by an amount that is unique to the token thats being hashed
-				bytes = GetBytes().Take(quantityIndexBytes).ToList();
-				if (bytes.Count == 3) { bytes.Add(0); }
-				yield return Math.Abs(BitConverter.ToInt32(bytes.ToArray(), 0)) % maxIndex;
-			}
-			yield break;
-		}
-
+							
 		public byte[] GetBytes(int quantity)
 		{
 			return GetBytes().Take(quantity).ToArray();
@@ -100,7 +47,7 @@ namespace BloomFilterCore
 			yield break;
 		}
 
-		private void Shuffle(int rounds)
+		protected void Shuffle(int rounds)
 		{
 			if (IsDisposed) { throw new ObjectDisposedException(this.GetType().Name); }
 
@@ -148,7 +95,7 @@ namespace BloomFilterCore
 			_table[j] = l;
 		}
 
-		private void Clear()
+		protected void Clear()
 		{
 			i = 0;
 			j = 0;
